requirements.txt
pytest==8.4.1
pytest-cov==7.0.0
PyYAML==6.0.3

